! 환경 설정 정리

+++ manage.py
@@ -41,7 +41,7 @@ def exec_command_line(exe_path, args):
 
 def install_package(ns):
     if not ns.package_names:
-        print 'NO_SOURCE_PATH'
+        print 'NO_PACKAGE_NAME'
         return -101
 
     for package_name in ns.package_names:
@@ -49,24 +49,36 @@ def install_package(ns):
 
     exec_command_line('pip', ['freeze', '> requirements.txt'])
 
+def reset_all_databases(ns):
+    print '#### reset all databases'
+    print '* config_path: %s' % ns.config_path
+    print '* database uri: %s' % server.app.config['SQLALCHEMY_DATABASE_URI']
+    for key, value in sorted(server.app.config['SQLALCHEMY_BINDS'].iteritems()):
+        print ' * bind_key: %s uri:%s' % (key, value)
+
+    print '* reset_all_password:',
+    password = raw_input()
+    if password != server.app.config['RESET_ALL_PASSWORD']:
+        print 'WRONG_DROP_ALL_PASSWORD'
+        return -102
+
+    server.db.drop_all() 
+    server.db.create_all()
+
 def run_script(ns):
     if not ns.source_paths:
         print 'NO_SOURCE_PATH'
-        return -101
+        return -103
 
     for source_path in ns.source_paths:
         execfile(source_path, globals())
 
 def run_shell(ns):
-    server.env.load_config_file(ns.config_path)
     code.interact('SHELL', local=dict(server=server))
 
 def run_server(ns):
-    server.env.load_config_file(ns.config_path)
+    server.env.prepare_all()
 
-    server.db.drop_all()
-    server.db.create_all()
-  
     server.app.register_blueprint(server.blog.bp)
     server.app.run('0.0.0.0', port=ns.port)
 
@@ -80,6 +92,9 @@ def main(program_path, program_args):
     install_package_parser = sub_parsers.add_parser('install_package')
     install_package_parser.add_argument('package_names', type=str, nargs='+', help='package names')
     install_package_parser.set_defaults(func=install_package)
+
+    reset_all_databases_parser = sub_parsers.add_parser('reset_all_databases')
+    reset_all_databases_parser.set_defaults(func=reset_all_databases)
 
     run_script_parser = sub_parsers.add_parser('run_script')
     run_script_parser.add_argument('source_paths', type=str, nargs='+', help='python source path') 
@@ -97,6 +112,7 @@ def main(program_path, program_args):
         return -1
 
     ns = main_parser.parse_args(program_args)
+    server.env.load_config_file(ns.config_path)
     return ns.func(ns)
 
 if __name__ == '__main__':
